Add #5 - controller
컨트롤러 제작
