+++ src/main/java/com/breadmore/coronaline/controller/PlaceController.java
@@ -5,25 +5,18 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-@RequestMapping("/admin")
+@RequestMapping("/places")
 @Controller
 public class PlaceController {
 
-    @GetMapping("/places")
-    public String adminPlaces() {
-        return "admin/places";
+    @GetMapping
+    public String places() {
+        return "place/index";
     }
-    @GetMapping("/places/{placeId}")
-    public String adminPlaceDetail(@PathVariable Integer placeId) {
-        return "admin/places-detail";
-    }
-    @GetMapping("/events")
-    public String adminEvents() {
-        return "admin/events";
-    }
-    @GetMapping("/events/{eventId}")
-    public String adminEventDetail(@PathVariable Integer eventId) {
-        return "admin/events-detail";
+
+    @GetMapping("/{placeId}")
+    public String placeDetail(@PathVariable Long placeId) {
+        return "place/detail";
     }
 
 }
